+++ Assignment2/Controllers/EmployeeController.cs
@@ -19,6 +19,7 @@ namespace Assignment2.Controllers
         [HttpGet]
         public async Task<ActionResult<ICollection<EmployeeDto>>> GetEmployeesByDepartmentAsync([FromRoute] string departmentId)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var employeesByDepartmentIdServiceResponse = await _employeeService.GetEmployeesByDepartmentIdAsync(departmentId);
             return employeesByDepartmentIdServiceResponse.IsSuccess ? Ok(employeesByDepartmentIdServiceResponse.Data) : NotFound(employeesByDepartmentIdServiceResponse.Message);
         }
@@ -26,6 +27,7 @@ namespace Assignment2.Controllers
         [HttpGet("{employeeId}")]
         public async Task<ActionResult<EmployeeDto>> GetEmployeeByIdAsync([FromRoute] string departmentId, [FromRoute] string employeeId)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var employeeByIdServiceResponse = await _employeeService.GetEmployeeByIdAsync(departmentId, employeeId);
             return (employeeByIdServiceResponse.IsSuccess) ? Ok(employeeByIdServiceResponse.Data) : NotFound(employeeByIdServiceResponse.Message);
         }
@@ -49,6 +51,7 @@ namespace Assignment2.Controllers
         [HttpDelete("{employeeId}")]
         public async Task<IActionResult> DeleteEmployeeByIdAsync([FromRoute] string departmentId, [FromRoute] string employeeId)
         {
+            if (!ModelState.IsValid) return BadRequest();
             var isEmployeeDeletedServiceResponse = await _employeeService.DeleteEmployeeByIdAsync(departmentId, employeeId);
             return isEmployeeDeletedServiceResponse.IsSuccess ? Ok() : NotFound(isEmployeeDeletedServiceResponse.Message);
         }
